# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Sample of WebApiRest_Golang",
        "title": "Golang Gin API",
        "termsOfService": "https://github1.com/CastroEduardo/golang-api-rest",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://github1.com/CastroEduardo/golang-api-rest/blob/master/LICENSE"
        },
        "version": "2.0"
    },
    "paths": {
...
        "/api/v2/addusersys": {
            "post": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Api-v2"
                ],
                "summary": "Add user to System",
                "parameters": [
                    {
                        "type": "string",
                        "description": "modelUser",
                        "name": "modelUser",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v2/manageDepartament": {
            "post": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Api-v2"
                ],
                "summary": "Manage Departament Sys",
                "parameters": [
                    {
                        "type": "string",
                        "description": "typeOperation",
                        "name": "typeOperation",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "idParam",
                        "name": "idParam",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "modelJson",
                        "name": "modelJson",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/api/v2/userlistsys": {
            "post": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Api-v2"
                ],
                "summary": "Get list userSys",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "app.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "bearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}